rename to common parser, add checkbox
