+++ src/app/core/models/templates/team-response-template.model.ts
@@ -2,7 +2,7 @@ import { IssueComment } from '../comment.model';
 import { GithubComment } from '../github/github-comment.model';
 import { DuplicateOfSection } from './sections/duplicate-of-section.model';
 import { Section } from './sections/section.model';
-import { buildTeamResponseSectionParser } from './sections/team-response-section-parser.model';
+import { buildTeamResponseSectionParser } from './sections/common-parsers.model';
 import { Header, Template } from './template.model';
 
 const { choice, coroutine, digits, str, whitespace } = require('arcsecond');
