+++ src/app/core/models/templates/tutor-moderation-issue-template.model.ts
@@ -3,7 +3,7 @@ import { IssueDispute } from '../issue-dispute.model';
 import { IssueDisputeSectionParser } from './sections/issue-dispute-section-parser.model';
 import { IssueDisputeSection } from './sections/issue-dispute-section.model';
 import { Section } from './sections/section.model';
-import { buildTeamResponseSectionParser } from './sections/team-response-section-parser.model';
+import { buildTeamResponseSectionParser } from './sections/common-parsers.model';
 import { Header, Template } from './template.model';
 
 const { coroutine, everyCharUntil, many1, str, whitespace } = require('arcsecond');
